Change variable types

+++ src/map.rs
@@ -66,13 +66,13 @@ fn update_enemy_checkpoint(
 fn damage_base(
   commands: &mut Commands,
   entity: &Entity,
-  enemy_health: i32,
+  enemy_health: u32,
   base: &mut Base
 ) {
   commands.entity(*entity).despawn_recursive();
   
   if base.health > 0 {
-    base.health -= enemy_health;
+    base.health -= enemy_health as i32;
   }
   
   if base.health <= 0{
